Move asset trade activity code to new class
