+++ core/src/main/java/bisq/core/app/BisqSetup.java
@@ -49,6 +49,7 @@ import bisq.core.provider.fee.FeeService;
 import bisq.core.provider.price.PriceFeedService;
 import bisq.core.trade.Trade;
 import bisq.core.trade.TradeManager;
+import bisq.core.trade.statistics.AssetTradeActivityCheck;
 import bisq.core.trade.statistics.TradeStatisticsManager;
 import bisq.core.user.Preferences;
 import bisq.core.user.User;
@@ -150,6 +151,7 @@ public class BisqSetup {
     private final PriceAlert priceAlert;
     private final MarketAlerts marketAlerts;
     private final VoteResultService voteResultService;
+    private final AssetTradeActivityCheck tradeActivityCheck;
     private final BSFormatter formatter;
     @Setter
     @Nullable
@@ -223,6 +225,7 @@ public class BisqSetup {
                      PriceAlert priceAlert,
                      MarketAlerts marketAlerts,
                      VoteResultService voteResultService,
+                     AssetTradeActivityCheck tradeActivityCheck,
                      BSFormatter formatter) {
 
 
@@ -259,6 +262,7 @@ public class BisqSetup {
         this.priceAlert = priceAlert;
         this.marketAlerts = marketAlerts;
         this.voteResultService = voteResultService;
+        this.tradeActivityCheck = tradeActivityCheck;
         this.formatter = formatter;
     }
 
@@ -624,6 +628,7 @@ public class BisqSetup {
         }
 
         tradeStatisticsManager.onAllServicesInitialized();
+        tradeActivityCheck.onAllServicesInitialized();
 
         accountAgeWitnessService.onAllServicesInitialized();
