+++ core/src/main/java/bisq/core/dao/governance/asset/AssetService.java
@@ -18,7 +18,6 @@
 package bisq.core.dao.governance.asset;
 
 import bisq.core.app.BisqEnvironment;
-import bisq.core.dao.state.DaoStateService;
 import bisq.core.locale.CurrencyUtil;
 
 import bisq.common.proto.persistable.PersistedDataHost;
@@ -34,10 +33,7 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 public class AssetService implements PersistedDataHost {
-
-
     private final Storage<RemovedAssetsList> storage;
-    private final DaoStateService daoStateService;
     @Getter
     private final RemovedAssetsList removedAssetsList = new RemovedAssetsList();
 
@@ -47,9 +43,8 @@ public class AssetService implements PersistedDataHost {
     ///////////////////////////////////////////////////////////////////////////////////////////
 
     @Inject
-    public AssetService(Storage<RemovedAssetsList> storage, DaoStateService daoStateService) {
+    public AssetService(Storage<RemovedAssetsList> storage) {
         this.storage = storage;
-        this.daoStateService = daoStateService;
     }
