+++ core/src/main/java/bisq/core/trade/statistics/TradeStatisticsManager.java
@@ -18,8 +18,6 @@
 package bisq.core.trade.statistics;
 
 import bisq.core.app.AppOptionKeys;
-import bisq.core.dao.governance.asset.AssetService;
-import bisq.core.locale.CryptoCurrency;
 import bisq.core.locale.CurrencyTuple;
 import bisq.core.locale.CurrencyUtil;
 import bisq.core.locale.Res;
@@ -34,33 +32,21 @@ import bisq.network.p2p.storage.persistence.AppendOnlyDataStoreService;
 import bisq.common.UserThread;
 import bisq.common.storage.JsonFileManager;
 import bisq.common.storage.Storage;
-import bisq.common.util.Tuple2;
 import bisq.common.util.Utilities;
-
-import org.bitcoinj.core.Coin;
 
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
 
-import com.google.common.base.Joiner;
-
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableSet;
-
-import java.time.Duration;
-
-import java.text.SimpleDateFormat;
 
 import java.io.File;
 
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
@@ -75,7 +61,6 @@ public class TradeStatisticsManager {
     private final P2PService p2PService;
     private final PriceFeedService priceFeedService;
     private final ReferralIdService referralIdService;
-    private final AssetService assetService;
     private final boolean dumpStatistics;
     private final ObservableSet<TradeStatistics2> observableTradeStatisticsSet = FXCollections.observableSet();
 
@@ -85,13 +70,11 @@ public class TradeStatisticsManager {
                                   TradeStatistics2StorageService tradeStatistics2StorageService,
                                   AppendOnlyDataStoreService appendOnlyDataStoreService,
                                   ReferralIdService referralIdService,
-                                  AssetService assetService,
                                   @Named(Storage.STORAGE_DIR) File storageDir,
                                   @Named(AppOptionKeys.DUMP_STATISTICS) boolean dumpStatistics) {
         this.p2PService = p2PService;
         this.priceFeedService = priceFeedService;
         this.referralIdService = referralIdService;
-        this.assetService = assetService;
         this.dumpStatistics = dumpStatistics;
         jsonFileManager = new JsonFileManager(storageDir);
 
@@ -100,14 +83,14 @@ public class TradeStatisticsManager {
 
     public void onAllServicesInitialized() {
         if (dumpStatistics) {
-            ArrayList<CurrencyTuple> fiatCurrencyList = new ArrayList<>(CurrencyUtil.getAllSortedFiatCurrencies().stream()
+            ArrayList<CurrencyTuple> fiatCurrencyList = CurrencyUtil.getAllSortedFiatCurrencies().stream()
                     .map(e -> new CurrencyTuple(e.getCode(), e.getName(), 8))
-                    .collect(Collectors.toList()));
+                    .collect(Collectors.toCollection(ArrayList::new));
             jsonFileManager.writeToDisc(Utilities.objectToJson(fiatCurrencyList), "fiat_currency_list");
 
-            ArrayList<CurrencyTuple> cryptoCurrencyList = new ArrayList<>(CurrencyUtil.getAllSortedCryptoCurrencies().stream()
+            ArrayList<CurrencyTuple> cryptoCurrencyList = CurrencyUtil.getAllSortedCryptoCurrencies().stream()
                     .map(e -> new CurrencyTuple(e.getCode(), e.getName(), 8))
-                    .collect(Collectors.toList()));
+                    .collect(Collectors.toCollection(ArrayList::new));
             cryptoCurrencyList.add(0, new CurrencyTuple(Res.getBaseCurrencyCode(), Res.getBaseCurrencyName(), 8));
             jsonFileManager.writeToDisc(Utilities.objectToJson(cryptoCurrencyList), "crypto_currency_list");
         }
@@ -124,9 +107,8 @@ public class TradeStatisticsManager {
         observableTradeStatisticsSet.addAll(map.values());
 
         priceFeedService.applyLatestBisqMarketPrice(observableTradeStatisticsSet);
-        dump();
 
-        checkTradeActivity();
+        dump();
     }
 
     public void publishTradeStatistics(List<Trade> trades) {
@@ -161,7 +143,11 @@ public class TradeStatisticsManager {
         }
     }
 
-    public void addToMap(TradeStatistics2 tradeStatistics, boolean storeLocally) {
+    public ObservableSet<TradeStatistics2> getObservableTradeStatisticsSet() {
+        return observableTradeStatisticsSet;
+    }
+
+    private void addToMap(TradeStatistics2 tradeStatistics, boolean storeLocally) {
         if (!observableTradeStatisticsSet.contains(tradeStatistics)) {
             boolean itemAlreadyAdded = observableTradeStatisticsSet.stream()
                     .anyMatch(e -> (e.getOfferId().equals(tradeStatistics.getOfferId())));
@@ -177,14 +163,10 @@ public class TradeStatisticsManager {
         }
     }
 
-    public void addToMap(TradeStatistics2 tradeStatistics, Map<String, TradeStatistics2> map) {
+    private void addToMap(TradeStatistics2 tradeStatistics, Map<String, TradeStatistics2> map) {
         TradeStatistics2 prevValue = map.putIfAbsent(tradeStatistics.getOfferId(), tradeStatistics);
         if (prevValue != null)
             log.debug("We have already an item with the same offer ID. That might happen if both the maker and the taker published the tradeStatistics");
-    }
-
-    public ObservableSet<TradeStatistics2> getObservableTradeStatisticsSet() {
-        return observableTradeStatisticsSet;
     }
 
     private void dump() {
@@ -201,149 +183,5 @@ public class TradeStatisticsManager {
             list.toArray(array);
             jsonFileManager.writeToDisc(Utilities.objectToJson(array), "trade_statistics");
         }
-    }
-
-    private void checkTradeActivity() {
-        Date compareDate = new Date(new Date().getTime() - Duration.ofDays(120).toMillis());
-        long minTradeAmount = Coin.parseCoin("0.001").value;
-        long minNumOfTrades = 3;
-
-        Map<String, Tuple2<Long, Integer>> tradeStatMap = new HashMap<>();
-        observableTradeStatisticsSet.stream()
-                .filter(e -> CurrencyUtil.isCryptoCurrency(e.getBaseCurrency()))
-                .filter(e -> e.getTradeDate().getTime() > compareDate.getTime())
-                .forEach(e -> {
-                    tradeStatMap.putIfAbsent(e.getBaseCurrency(), new Tuple2<>(0L, 0));
-                    Tuple2<Long, Integer> tuple2 = tradeStatMap.get(e.getBaseCurrency());
-                    long accumulatedTradeAmount = tuple2.first + e.getTradeAmount().getValue();
-                    int numTrades = tuple2.second + 1;
-                    tradeStatMap.put(e.getBaseCurrency(), new Tuple2<>(accumulatedTradeAmount, numTrades));
-                });
-        StringBuilder newAssets = new StringBuilder("\nNew assets (in warming up phase):");
-        StringBuilder sufficientlyTraded = new StringBuilder("\nSufficiently traded assets:");
-        StringBuilder insufficientlyTraded = new StringBuilder("\nInsufficiently traded assets:");
-        StringBuilder notTraded = new StringBuilder("\nNot traded assets:");
-        List<CryptoCurrency> whiteListedSortedCryptoCurrencies = CurrencyUtil.getWhiteListedSortedCryptoCurrencies(assetService);
-        Set<CryptoCurrency> assetsToRemove = new HashSet<>(whiteListedSortedCryptoCurrencies);
-        whiteListedSortedCryptoCurrencies.forEach(e -> {
-            String code = e.getCode();
-            String nameAndCode = CurrencyUtil.getNameAndCode(code);
-            long tradeAmount = 0;
-            int numTrades = 0;
-            boolean isInTradeStatMap = tradeStatMap.containsKey(code);
-            if (isInTradeStatMap) {
-                Tuple2<Long, Integer> tuple = tradeStatMap.get(code);
-                tradeAmount = tuple.first;
-                numTrades = tuple.second;
-            }
-            if (isWarmingUp(code)) {
-                assetsToRemove.remove(e);
-                newAssets.append("\n")
-                        .append(nameAndCode)
-                        .append(": Trade amount: ")
-                        .append(Coin.valueOf(tradeAmount).toFriendlyString())
-                        .append(", number of trades: ")
-                        .append(numTrades);
-            }
-            if (!isWarmingUp(code) && !hasPaidBSQFee(code)) {
-                if (isInTradeStatMap) {
-                    if (tradeAmount >= minTradeAmount || numTrades >= minNumOfTrades) {
-                        assetsToRemove.remove(e);
-                        sufficientlyTraded.append("\n")
-                                .append(nameAndCode)
-                                .append(": Trade amount: ")
-                                .append(Coin.valueOf(tradeAmount).toFriendlyString())
-                                .append(", number of trades: ")
-                                .append(numTrades);
-                    } else {
-                        insufficientlyTraded.append("\n")
-                                .append(nameAndCode)
-                                .append(": Trade amount: ")
-                                .append(Coin.valueOf(tradeAmount).toFriendlyString())
-                                .append(", number of trades: ")
-                                .append(numTrades);
-                    }
-                } else {
-                    notTraded.append("\n").append(nameAndCode);
-                }
-            }
-        });
-        List<CryptoCurrency> assetsToRemoveList = assetsToRemove.stream()
-                .sorted(Comparator.comparing(CryptoCurrency::getCode))
-                .collect(Collectors.toList());
-
-        String result = "Date for checking trade activity: " + new SimpleDateFormat("yyyy-MM-dd'T'").format(compareDate) +
-                "\n\nAssets to remove (" + assetsToRemoveList.size() + "):\n" + Joiner.on("\n").join(assetsToRemoveList) +
-                "\n\n" + insufficientlyTraded.toString() +
-                "\n\n" + notTraded.toString() +
-                "\n\n" + newAssets.toString() +
-                "\n\n" + sufficientlyTraded.toString();
-        // Utilities.copyToClipboard(result);
-        log.debug(result);
-    }
-
-    private boolean hasPaidBSQFee(String code) {
-        return assetService.hasPaidBSQFee(code);
-    }
-
-    private boolean isWarmingUp(String code) {
-        Set<String> newlyAdded = new HashSet<>();
-
-        // v0.7.1 Jul 4 2018
-        newlyAdded.add("ZOC");
-        newlyAdded.add("AQUA");
-        newlyAdded.add("BTDX");
-        newlyAdded.add("BTCC");
-        newlyAdded.add("BTI");
-        newlyAdded.add("CRDS");
-        newlyAdded.add("CNMC");
-        newlyAdded.add("TARI");
-        newlyAdded.add("DAC");
-        newlyAdded.add("DRIP");
-        newlyAdded.add("FTO");
-        newlyAdded.add("GRFT");
-        newlyAdded.add("LIKE");
-        newlyAdded.add("LOBS");
-        newlyAdded.add("MAX");
-        newlyAdded.add("MEC");
-        newlyAdded.add("MCC");
-        newlyAdded.add("XMN");
-        newlyAdded.add("XMY");
-        newlyAdded.add("NANO");
-        newlyAdded.add("NPW");
-        newlyAdded.add("NIM");
-        newlyAdded.add("PIX");
-        newlyAdded.add("PXL");
-        newlyAdded.add("PRIV");
-        newlyAdded.add("TRIT");
-        newlyAdded.add("WAVI");
-
-        // v0.8.0 Aug 22 2018
-        // none added
-
-        // v0.9.0 (Date TBD)
-        newlyAdded.add("ACM");
-        newlyAdded.add("BTC2");
-        newlyAdded.add("BLUR");
-        newlyAdded.add("CHA");
-        newlyAdded.add("CROAT");
-        newlyAdded.add("DRGL");
-        newlyAdded.add("ETHS");
-        newlyAdded.add("GBK");
-        newlyAdded.add("KEK");
-        newlyAdded.add("LOKI");
-        newlyAdded.add("MBGL");
-        newlyAdded.add("NEOS");
-        newlyAdded.add("PZDC");
-        newlyAdded.add("QMCoin");
-        newlyAdded.add("QRL");
-        newlyAdded.add("RADS");
-        newlyAdded.add("RYO");
-        newlyAdded.add("SUB1X");
-        newlyAdded.add("MAI");
-        newlyAdded.add("TRTL");
-        newlyAdded.add("ZER");
-
-        return newlyAdded.contains(code);
     }
 }
